refactor(profile): extract the profile domain core — D8 move
The public profile's shape was owned by the editor's wizard. `app/[slug]/page.tsx`
had to reach into `app/dashboard/profile/` to find out what a profile IS — the
D-4 privacy finding, and three of the five wrong-direction app→app edges.

**app→app edges: 5 → 2.** All three `[slug] → dashboard/profile` edges are gone;
the two that remain are pre-existing and unrelated (`dashboard → (auth)`, and
`(legal)/about → _marketing`, which is already on KAN-422's DELETE list).

Moved into `src/modules/profile/`: profile-fields.ts, section-visibility.ts,
manual-of-me-fields.ts, favourites.ts, visibility.ts, and country-codes.ts
(out of `lib/affiliate/`, per the plan). 26 importers repointed.

⚠️ THE PLAN SAID "MOVE types.tsx". IT CANNOT BE MOVED — IT MUST BE SPLIT.
That file is not a domain file. It is seven domain interfaces welded to two
React components carrying founder-owned design tokens (`--color-sage`,
`--color-ink`), and `profile` is `uiApprovalGated: always`. Moving it wholesale
would drag a button into a domain module that everything else is meant to depend
on freely. So the interfaces moved to `src/modules/profile/types.ts` and `Field`
/ `SaveButton` stayed in the app tree. 124 → 54 lines.

`steps/types.tsx` re-exports the types, so all 14 existing editor imports are
untouched: this commit moves the boundary, it does not rewrite the wizard. New
code should import from the module.

Caught by the estate's own guards, all four of them mine:
* THE SRC MANIFEST TRAP AGAIN — three keys (favourites, profileFields,
  sectionVisibility) were silently DROPPED, read only via SRC and hard-coded
  nowhere that counts. Two went `undefined`, which makes an assertion pass
  forever. source-path-manifest-integrity caught it; seeds added.
* CTL-041 flagged the count twice. The first derivation was WRONG because I ran
  it before staging — `git ls-files` cannot see an untracked file, which is the
  documented "stage before measuring" gotcha, hit again.
* The extraction DoD's stale-ref rule would have flagged three COMMENTS naming
  moved paths; fixed at source rather than suppressed.
* affiliate's own count fell 8 → 7, derived, not assumed.

Not in this commit: D-6 (removing `is_published` from ALLOWED_PROFILE_FIELDS)
is a BEHAVIOUR change and lands separately, so this one stays reviewable as a
pure move.

277 suites / 3619 tests, tsc clean, eslint 0 errors, dependency gate 0 errors.

Docs-Updated: KAN-415 ARCHITECTURE.md gains the profile domain-core section and
why the UI/domain split is a privacy boundary rather than tidiness.
UI-Change-Approved: KAN-415
