+++ src/app/dashboard/profile/page.tsx
@@ -1,8 +1,8 @@
 import { createClient } from '@/modules/platform/supabase-server';
 import { redirect } from 'next/navigation';
 import { EditProfileForm } from './edit-profile-form';
-import type { ManualOfMe } from './manual-of-me-fields';
-import { MANUAL_OF_ME_FIELDS } from './manual-of-me-fields';
+import type { ManualOfMe } from '@/modules/profile/manual-of-me-fields';
+import { MANUAL_OF_ME_FIELDS } from '@/modules/profile/manual-of-me-fields';
 import { isConveneEnabledForCurrentUser } from '@/lib/convene/flags-user';
 import { getRecommendations } from '@/lib/recommend';
 import { keyForRecommendation } from '@/lib/recommend/dismissals';
