+++ tests/support/source-paths.json
@@ -1,7 +1,7 @@
 {
   "$comment": "GENERATED — regenerate with `npm run gen:test-paths`. Mirror of tests/support/source-paths.ts for .cjs and shell consumers (KAN-414 F4).",
   "generated_from": "literals referenced by tracked files under tests/",
-  "count": 194,
+  "count": 195,
   "SRC": {
     "codeowners": ".github/CODEOWNERS",
     "pullRequestTemplate": ".github/PULL_REQUEST_TEMPLATE.md",
@@ -110,13 +110,10 @@
     "conversationStartersActions": "src/app/dashboard/profile/conversation-starters-actions.ts",
     "deliveryCountryActions": "src/app/dashboard/profile/delivery-country-actions.ts",
     "editProfileForm": "src/app/dashboard/profile/edit-profile-form.tsx",
-    "favourites": "src/app/dashboard/profile/favourites.ts",
     "filesActions": "src/app/dashboard/profile/files-actions.ts",
     "legacyPage": "src/app/dashboard/profile/legacy/page.tsx",
     "manualOfMeActions": "src/app/dashboard/profile/manual-of-me-actions.ts",
     "profilePage": "src/app/dashboard/profile/page.tsx",
-    "profileFields": "src/app/dashboard/profile/profile-fields.ts",
-    "sectionVisibility": "src/app/dashboard/profile/section-visibility.ts",
     "profileSections": "src/app/dashboard/profile/sections",
     "steps": "src/app/dashboard/profile/steps",
     "conversationStartersStep": "src/app/dashboard/profile/steps/conversation-starters-step.tsx",
@@ -181,6 +178,10 @@
     "platformEnv": "src/modules/platform/env.ts",
     "supabaseServer": "src/modules/platform/supabase-server.ts",
     "supabaseService": "src/modules/platform/supabase-service.ts",
+    "favourites": "src/modules/profile/favourites.ts",
+    "profileFields": "src/modules/profile/profile-fields.ts",
+    "sectionVisibility": "src/modules/profile/section-visibility.ts",
+    "profileTypes": "src/modules/profile/types.ts",
     "database": "src/types/database",
     "index": "src/types/database/index.ts",
     "prod": "src/types/database/prod.ts",
