+++ tests/unit/affiliations-and-autosave.test.ts
@@ -361,7 +361,7 @@ describe('KAN-220: surface-area regression guards', () => {
 
   test('WizardSchool type declares affiliation_type', () => {
     const src = readFileSync(
-      resolve(ROOT, SRC.types),
+      resolve(ROOT, SRC.profileTypes),
       'utf-8',
     );
     // Type now carries affiliation_type — KAN-220 schools/orgs/communities split
