+++ tests/unit/kan444-custom-group-persistence.test.ts
@@ -66,7 +66,7 @@ jest.mock('@/modules/platform/supabase-server', () => ({
 }));
 
 import { addProfileItem } from '@/app/dashboard/profile/actions';
-import { CUSTOM_FAVOURITE_CATEGORY } from '@/app/dashboard/profile/favourites';
+import { CUSTOM_FAVOURITE_CATEGORY } from '@/modules/profile/favourites';
 
 beforeEach(() => {
   mockInsertCapture.mockClear();
