+++ tests/unit/kan444-favourites-groups.test.ts
@@ -61,7 +61,7 @@ import {
   favouriteLabelForItem,
   isFavouriteCategory,
   primaryCategory,
-} from '@/app/dashboard/profile/favourites';
+} from '@/modules/profile/favourites';
 
 const ROOT = resolve(__dirname, '../..');
 
@@ -376,7 +376,7 @@ describe('KAN-444: add-your-own — a group the member names', () => {
 describe('KAN-444: both surfaces render from the shared grouping', () => {
   test('the public profile groups via groupFavourites and keeps no list of its own', () => {
     const page = readSource(SRC.slugPage);
-    expect(page).toContain("from '@/app/dashboard/profile/favourites'");
+    expect(page).toContain("from '@/modules/profile/favourites'");
     expect(page).toContain('groupFavourites(typedItems)');
     // The old per-surface table is gone; if a future change re-hardcodes the
     // categories here, the two surfaces can drift again.
@@ -387,7 +387,7 @@ describe('KAN-444: both surfaces render from the shared grouping', () => {
 
   test('the editor favourites section takes its categories and picker from the module', () => {
     const editor = readSource(SRC.editProfileForm);
-    expect(editor).toContain("from './favourites'");
+    expect(editor).toContain('@/modules/profile/favourites');
     expect(editor).toMatch(/id:\s*'favourites'[\s\S]{0,500}?categories:\s*FAVOURITE_CATEGORIES/);
     expect(editor).toMatch(/categoryOptions:\s*FAVOURITE_CATEGORY_OPTIONS/);
     expect(editor).toMatch(/groupLabelCategory:\s*CUSTOM_FAVOURITE_CATEGORY/);
