+++ tests/unit/section-visibility.test.ts
@@ -29,7 +29,7 @@ import {
   getEffectiveItemVisibility,
   isItemVisibleUnderHybridModel,
   type SectionVisibility,
-} from '@/app/dashboard/profile/section-visibility';
+} from '@/modules/profile/section-visibility';
 
 describe('KAN-221: CONTROLLABLE_SECTION_KEYS + ITEM_CATEGORY_TO_SECTION', () => {
   test('every controllable section key has at least one item category mapped to it', () => {
